Removed title from mcon playerAnimator image page Updated parchment and rei Made power_color field in Power optional, defaults to gray Made cookingTime in TeapotBrewingRecipe intprovider, allowing for random times in a range or condition dependent times Made teapot brewing time int Updated tea recipe example Updated none.json to remove color Fixed teapot not boiling in nether

+++ Common/src/main/java/dev/thomasglasser/minejago/world/level/block/component/TeapotBlockComponentProvider.java
@@ -54,7 +54,7 @@ public enum TeapotBlockComponentProvider implements IBlockComponentProvider {
                 iTooltip.append(Component.translatable("block.minejago.teapot.waila.cups", cups));
             }
 
-            short time = be.getBrewTime();
+            int time = be.getBrewTime();
             if (time > 0)
             {
                 IElement icon = elementHelper.item(Items.CLOCK.getDefaultInstance(), 0.5f).translate(new Vec2(0, -2));
